Don't require an API request to determine the type
Most of the events coming through feed.query right now are of
type CMIT, meaning we have an extra API request per commit only
to ignore it. Since "CMIT" is embeded in the PHID itself, just use that
and skip the request for most things we'll ignore.

Also add some logging.

Change-Id: If69a31d67ad00e8167aeeb4f2fbba651a017ca1a