Replace fabfile with command line script to run remotely
Fabric is too confusing to use. This replaces it by a 'manage.py'
script that has to be run /on/ tool labs.

Support for taxonomy.py has been removed (as this tool has been
malfunctioning for over a year, and no-one complained).

Updating contributors was moved to a separate tools/ script.

Bug: T201958
Change-Id: I4de020f209edac792e8608a49d242a7585de069c