add ServiceOps -> #wikimedia-serviceops
A new IRC channel has been created in T211902
for the Service Operations team.

A new Phabricator tag for the team has been
requested in T209522.

Map these together to get notifications on the
channel for tickets with the new tag.

A request has been made in #freenode-sigyn to get
an exempt for wikibugs for the new channel in Sigyn.

Please note both the tag and the exempt are pending
at the time of this writing, so merge needs to wait
for them.

Bug: T211912
Change-Id: I473eab166620d1218b108ae7917c37a03a5d499a