T241109: Throw exception when Phabricator returns an error
This will cause the bot to crash and restart, which will cause the
polling cycle to start from scratch as well. This is not the cleanest
possible implementation but does lay the ground work for further
(cleaner) implementation work, by catching and handling the exception.

For now, the idea is to run and monitor this to see if we see any other
PhabricatorExceptions popping up, so we can also handle those.

The processing order is now explicitly old-to-new as the mocked requests
do not necessarily retain their sorted order, and it is preferable not to
depend on the order of values in a JSON object.

Bug: T241109
Change-Id: I6870a9f83508113a60ccaeab6e2467ae77916892
Reviewed-on: https://gerrit.wikimedia.org/r/561294
Reviewed-by: Merlijn van Deen <valhallasw@arctus.nl>
Tested-by: jenkins-bot