releng: Remove tags RelEng team doesn't own
The signal/noise level of -releng is getting to be too noisy. We need to
reduce the bot activity to only projects owned/maintained by the Release
Engineering team. This is in service to us closing our private IRC team
channel and using the -releng channel for project conversations.
Currently it is not feasible with the level of bot activity.

Change-Id: Ib030a3a9bb54cf9f7cd525c9e6bc6919f05c2be1
Reviewed-on: https://gerrit.wikimedia.org/r/c/labs/tools/wikibugs2/+/704636
Reviewed-by: Jforrester <jforrester@wikimedia.org>
Tested-by: jenkins-bot