Rename jerkins-bot to 'CI reject'
CI reports to gerrit as "jenkins-bot", if I remember properly some had a
notification setup for "jenkins" and thus kept being poked for every
failure. As a pun, it got renamed jerkins.

It is a bit meaningless since it is not a real user and I find the pun a
bit odd after all those years.

The new name detaches from a username and makes the overall message a
bit more meaningful.

Update tests to expect user 'CI reject'.

Change-Id: I87f4f36165ace4e60910a9f7ac0ddbeef9fb18c5
Reviewed-on: https://gerrit.wikimedia.org/r/c/labs/tools/wikibugs2/+/791694
Reviewed-by: Samtar <starling-ctr@wikimedia.org>
Reviewed-by: Jforrester <jforrester@wikimedia.org>
Tested-by: jenkins-bot