Remove unrelated projects from #-fundraising
The previous regex was letting a lot of unrelated projects through
(basically anything with "Fundraising" in the title). The regex did
attempt to prevent this by doing explicit ignores on a couple false
positives, but the explicit ignore does not scale.

I've instead made it an explicit allow. All of our projects follow one
of two patterns:
* Fundraising.*
* Wikimedia-Fundraising.*

Change-Id: If9779db5b66300346142c53418a110143e8485fd
Reviewed-on: https://gerrit.wikimedia.org/r/c/labs/tools/wikibugs2/+/859119
Reviewed-by: Jforrester <jforrester@wikimedia.org>
Tested-by: jenkins-bot