Remove #wikimedia-services
The channel has a single entity speaking, wikibugs. I tried to
understand if anyone is even remotely using it for anything, I got
nothing back.

RESTBase is a deprecated software/infrastructure, coordination for it's
eventual demise happens via Slack. I doubt this removal would somehow
impact involved stakeholders

service-runner is still alive, but out of all the people that have tagged or
interacted with a service-runner tagged task, in the last 2 years, only
3 are present in the channel. That would be kart_, ottomata and yours
truly. And we sync up in other ways and not via this channel

Change-Id: Ief9e50031137849a12be2c95816c7996a6a9fe9e
Reviewed-on: https://gerrit.wikimedia.org/r/c/labs/tools/wikibugs2/+/935728
Reviewed-by: Majavah <hi@taavi.wtf>
Tested-by: jenkins-bot