wikibugs: let logger handle message formatting
The debug/info/warning/etc logger methods treat the first argument as
a format string and subsequent unnamed arguments as values to be merged
into the format using the string formatting operator. The prior code we
had to do this manually is redundant and actually can cause extra work
at runtime as the default behavior will only expand the format string
when the log event is going to be emitted.

Change-Id: Ia78baf30e2bba4bf7722f349f7b6f8a860cdf386
Reviewed-on: https://gerrit.wikimedia.org/r/c/labs/tools/wikibugs2/+/1004316
Tested-by: jenkins-bot
Reviewed-by: Merlijn van Deen <valhallasw@arctus.nl>