test_wikibugs2: update test_add_project assertion
The test_add_project message url assertion is suddenly failing because
the code is now finding the deep link anchor for the tested transaction.
This must be a side effect of the recent 4b4ccb0 update to the
get_anchors_for_task business logic. Confusingly the test was passing
when that change was being reviewed and merged, but is failing now.
:shrug:

Bug: T199007
Change-Id: I5c3773cd29b7c021322ceec76cf1a0cf4c333e3e
Reviewed-on: https://gerrit.wikimedia.org/r/c/labs/tools/wikibugs2/+/1004314
Reviewed-by: Merlijn van Deen <valhallasw@arctus.nl>
Tested-by: jenkins-bot