wikibugs: Get anchors from API instead of screen scraping
Replace screen scraping with data found in the
`maniphest.gettasktransactions` API response.

This turns out to be a situation where we have to guess what transaction
id is actually used to link to a particular transaction group. We may
end up deciding that screen scraping was actually the best option here
as we really are interested in state that exists in the rendered HTML.

Bug: T1177
Change-Id: Ic13e6b978f5d2ebf76edd0785d555fc71f3f10aa
Reviewed-on: https://gerrit.wikimedia.org/r/c/labs/tools/wikibugs2/+/1004317
Reviewed-by: BryanDavis <bd808@wikimedia.org>
Tested-by: jenkins-bot